packages/config/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "computer-config"
version = "0.1.0"
description = "Site configuration SDK — typed access to site.yaml"
requires-python = ">=3.14"
dependencies = [
    "pydantic>=2.12.0",
    "pyyaml>=6.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["sdk"]

[tool.ruff]
line-length = 100
